Time: 1 ms (100.00%) | Memory: 42.5 MB (65.56%) - LeetSync

+++ 4168-mirror-distance-of-an-integer/mirror-distance-of-an-integer.java
@@ -1,6 +1,6 @@
 class Solution {
     public int mirrorDistance(int n) {
-        return (int) Math.abs(n - rev(n));
+        return Math.abs(n - rev(n));
     }
 
     private int rev(int n) {
@@ -9,6 +9,6 @@ class Solution {
             sum = sum * 10 + n % 10;
             n /= 10;
         }
-        return (int) sum;
+        return sum;
     }
 }
